adding header and source file to control file and folder, change all the library into "Requirement.h" and start to create linked list and folder function.
