+++ SFML/Components.h
@@ -155,114 +155,5 @@ public:
 //define namespace to work with files
 namespace fsys = std::filesystem;
 
-class File {//This class has only worked with CSV files
-private: 
-    std::fstream file;
-    std::vector<std::string> rowContent; //Path of the "mother" folder
-    std::string fileName;
-    fsys::path folderPath;
 
-public:
-    File() {}
 
-    File(fsys::path folderPath, std::string fileName) {
-        this->fileName = fileName;
-        this->folderPath = folderPath;
-        folderPath /= fileName;
-        std::fstream file(folderPath);
-    }
-
-    void open() {
-        this->file.open(this->folderPath / this->fileName);
-    }
-    
-    void close() {
-        this->file.close();
-    }
-
-    std::vector<std::string> readRow() {//Only store one row per call
-        this->rowContent.clear();
-        std::string row;
-        if (!(this->file >> row)) return this->rowContent;
-        std::getline(this->file, row, '\n');
-        std::stringstream sRow(row);
-        std::string element;
-        while (std::getline(sRow, element, ',')) {
-            this->rowContent.push_back(element);
-        }
-        return this->rowContent;
-    }
-
-    void write(std::string content) {
-        this->file << content << std::endl;
-    }
-
-    void update() {
-
-    }
-
-    ~File() {
-        if (file.is_open())
-            this->file.close();
-    }
-};
-
-class Folder {
-private:
-    fsys::path folderPath; //Path of this folder
-    std::string folderName;
-    bool CreateDirectoryRecursive(fsys::path const& folderDir, std::error_code& err)
-    {
-        err.clear();
-        if (!fsys::create_directories(folderDir, err))
-        {
-            if (fsys::exists(folderDir))
-            {
-                // The folder already exists:
-                err.clear();
-                return true;
-            }
-            return false;
-        }
-        return true;
-    }
-
-public:
-    Folder(std::string folderName) {
-        this->folderPath = "Data";
-        this->folderName = folderName;
-        std::error_code err;
-        if (!CreateDirectoryRecursive(this->folderPath, err))
-        {
-            // Report the error:
-            std::cout << "CreateDirectoryRecursive FAILED, err: " << err.message() << std::endl;
-        }
-    }
-
-    Folder(fsys::path folderPath, std::string folderName) {
-        this->folderPath = folderPath;
-        this->folderName = folderName;
-        std::error_code err;
-        if (!CreateDirectoryRecursive(this->folderPath, err))
-        {
-            // Report the error:
-            std::cout << "CreateDirectoryRecursive FAILED, err: " << err.message() << std::endl;
-        }
-    }
-
-    void createSubFolder(std::string folderName) {
-        Folder folder(this->folderPath, folderName);
-    }
-
-    void createSubFile(std::string fileName) {
-        File file(this->folderPath, fileName); //remember to use file.close() after
-    }
-
-    void ToSubFolder(std::string folderName) {
-        if (!fsys::exists(this->folderPath)) std::cerr << "No such directory " << this->folderPath << std::endl;
-        else {
-            this->folderPath /= this->folderName;
-            this->folderName = folderName;
-        }
-    }
-};
