+++ SFML/FolderFunction.cpp
@@ -1,0 +1,6 @@
+#include "FileNFolder.h"
+
+void create_multiple_directories(List <std::string> linkedlist) //TODO: Create multiple directories based on the linked list which a teacher input for classes in school year.
+{
+
+}
