+++ SFML/Requirement.h
@@ -7,6 +7,10 @@
 #include <string>
 #include <sstream>
 #include <filesystem>
+#include <string>
+#include <ctime>
+#include <iomanip>
+#include <sstream>
 #include "Drawio.h"
 #include "Activity.h"
 #include "LinkList.h"
