+++ SFML/SFML.vcxproj
@@ -104,10 +104,8 @@
       <SDLCheck>true</SDLCheck>
       <PreprocessorDefinitions>_DEBUG;_CONSOLE;%(PreprocessorDefinitions)</PreprocessorDefinitions>
       <ConformanceMode>true</ConformanceMode>
-
       <LanguageStandard>stdcpp17</LanguageStandard>
       <AdditionalIncludeDirectories>$(SolutionDir)\SFML_Lib\include</AdditionalIncludeDirectories>
-
     </ClCompile>
     <Link>
       <SubSystem>Console</SubSystem>
@@ -139,6 +137,7 @@
     <ClCompile Include="Components.cpp" />
     <ClCompile Include="Drawio.cpp" />
     <ClCompile Include="ExtraFunction.cpp" />
+    <ClCompile Include="FolderFunction.cpp" />
     <ClCompile Include="LinkList.cpp" />
     <ClCompile Include="Main.cpp" />
     <ClCompile Include="Student.cpp" />
@@ -150,6 +149,7 @@
     <ClInclude Include="Components.h" />
     <ClInclude Include="Drawio.h" />
     <ClInclude Include="ExtraFunction.h" />
+    <ClInclude Include="FileNFolder.h" />
     <ClInclude Include="LinkList.h" />
     <ClInclude Include="Requirement.h" />
     <ClInclude Include="resource.h" />
